# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Notification:
    properties:
      action_url:
        type: string
      archived_at:
        type: string
      comment:
        $ref: '#/definitions/models.TaskComment'
      comment_id:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      data:
        $ref: '#/definitions/models.NotificationData'
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      message:
        type: string
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      read_at:
        type: string
      status:
        $ref: '#/definitions/models.NotificationStatus'
      task:
        $ref: '#/definitions/models.Task'
      task_id:
        description: Related entity references
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.NotificationData:
    properties:
      actor_id:
        type: string
      actor_name:
        type: string
      entity_id:
        type: string
      entity_name:
        type: string
      entity_type:
        type: string
      extra_data:
        additionalProperties: true
        type: object
      new_value:
        type: string
      previous_value:
        type: string
    type: object
  models.NotificationStatus:
    enum:
    - unread
    - read
    - archived
    type: string
    x-enum-varnames:
    - NotificationStatusUnread
    - NotificationStatusRead
    - NotificationStatusArchived
  models.NotificationType:
    enum:
    - welcome
    - task_assigned
    - task_created
    - task_completed
    - task_updated
    - task_due
    - task_overdue
    - comment_added
    - user_invited
    - user_joined
    - project_created
    - system_update
    type: string
    x-enum-varnames:
    - NotificationTypeWelcome
    - NotificationTypeTaskAssigned
    - NotificationTypeTaskCreated
    - NotificationTypeTaskCompleted
    - NotificationTypeTaskUpdated
    - NotificationTypeTaskDue
    - NotificationTypeTaskOverdue
    - NotificationTypeCommentAdded
    - NotificationTypeUserInvited
    - NotificationTypeUserJoined
    - NotificationTypeProjectCreated
    - NotificationTypeSystemUpdate
  models.Project:
    properties:
      color:
        description: Hex color code
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      tasks:
        description: Relationships
        items:
          $ref: '#/definitions/models.Task'
        type: array
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.Tag:
    properties:
      color:
        description: Hex color code
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      name:
        type: string
      tasks:
        description: Relationships
        items:
          $ref: '#/definitions/models.Task'
        type: array
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.Task:
    properties:
      activities:
        items:
          $ref: '#/definitions/models.TaskActivity'
        type: array
      actual_hours:
        type: number
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: string
      attachments:
        items:
          $ref: '#/definitions/models.TaskAttachment'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.TaskComment'
        type: array
      completed_at:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creator_id:
        description: Relationships
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      parent:
        $ref: '#/definitions/models.Task'
      parent_id:
        type: string
      priority:
        $ref: '#/definitions/models.TaskPriority'
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      subtasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      title:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.TaskActivity:
    properties:
      action:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        type: string
      field:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      new_value:
        type: string
      old_value:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskAttachment:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      mime_type:
        type: string
      original_name:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskComment:
    properties:
      content:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      parent:
        $ref: '#/definitions/models.TaskComment'
      parent_id:
        type: string
      replies:
        items:
          $ref: '#/definitions/models.TaskComment'
        type: array
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskPriority:
    enum:
    - low
    - medium
    - high
    - urgent
    type: string
    x-enum-varnames:
    - TaskPriorityLow
    - TaskPriorityMedium
    - TaskPriorityHigh
    - TaskPriorityUrgent
  models.TaskStatus:
    enum:
    - todo
    - in_progress
    - in_review
    - completed
    - canceled
    type: string
    x-enum-varnames:
    - TaskStatusTodo
    - TaskStatusInProgress
    - TaskStatusInReview
    - TaskStatusCompleted
    - TaskStatusCanceled
...
  models.User:
    properties:
      assigned_tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      avatar:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      created_tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      default_task_priority:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      email_verified_at:
        type: string
      enable_email_notifications:
        type: boolean
      enable_push_notifications:
        type: boolean
      first_name:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      language:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      notifications:
        items:
          $ref: '#/definitions/models.Notification'
        type: array
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      show_completed_tasks:
        type: boolean
      status:
        $ref: '#/definitions/models.UserStatus'
      task_reminders:
        type: boolean
      task_view_mode:
        description: list, board, calendar
        type: string
      tasks_per_page:
        type: integer
      tenant:
        allOf:
        - $ref: '#/definitions/models.Tenant'
        description: Relationships
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      theme:
        description: User preferences (flattened)
        type: string
      timezone:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      weekly_digest:
        type: boolean
    type: object
...
  models.UserRole:
    enum:
    - admin
    - manager
    - user
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleManager
    - UserRoleUser
  models.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
    - UserStatusSuspended
...
  response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.PaginationResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@taskflow.com
    name: API Support
    url: http://www.taskflow.com/support
  description: A multi-tenant task management platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TaskFlow API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Get a paginated list of notifications for the current user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by read/unread status
        in: query
        name: status
        type: string
      - description: Filter by notification type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}:
    delete:
      description: Delete a specific notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete notification
      tags:
      - notifications
    get:
      description: Get a specific notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notification'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get notification
      tags:
      - notifications
  /notifications/{id}/read:
    put:
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /notifications/{id}/unread:
    put:
      description: Mark a specific notification as unread
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as unread
      tags:
      - notifications
  /notifications/mark-all-read:
    put:
      description: Mark all notifications as read for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /notifications/settings:
    get:
      description: Get notification preferences for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get notification settings
      tags:
      - notifications
    put:
      consumes:
      - application/json
      description: Update notification preferences for the current user
      parameters:
      - description: Notification preferences
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update notification settings
      tags:
      - notifications
  /notifications/unread-count:
    get:
      description: Get the count of unread notifications for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - notifications
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"